Seperating list to another file for testing sake
