+++ package.json
@@ -32,6 +32,7 @@
     "winston": "^3.8.2"
   },
   "devDependencies": {
-    "jest": "^29.5.0"
+    "jest": "^29.5.0",
+    "jsdom": "^22.0.0"
   }
 }
